api: move common code from http_endpoint into http
Move common code which will be shared v1/v2 API from http_endpoint.rs
into http.rs.
