+++ api/src/http.rs
@@ -4,31 +4,250 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use std::collections::HashMap;
-use std::io::{Error, ErrorKind, Result};
+use std::io::{self, Error, ErrorKind, Result};
 use std::os::unix::io::AsRawFd;
 use std::path::PathBuf;
-use std::sync::mpsc::{Receiver, Sender};
+use std::sync::mpsc::{Receiver, RecvError, SendError, Sender};
 use std::sync::Arc;
 use std::thread;
 use std::time::SystemTime;
 
+use dbs_uhttp::{Body, HttpServer, MediaType, Request, Response, ServerError, StatusCode, Version};
 use http::uri::Uri;
-use url::Url;
-
-use dbs_uhttp::{HttpServer, MediaType, Request, Response, ServerError, StatusCode};
 use mio::unix::SourceFd;
 use mio::{Events, Interest, Poll, Token, Waker};
+use nydus_utils::metrics::IoStatsError;
+use serde::Deserialize;
+use serde_json::Error as SerdeError;
+use url::Url;
 
 use crate::http_endpoint_v1::{
-    error_response, ApiError, ApiRequest, ApiResponse, EventsHandler, ExitHandler, FsBackendInfo,
-    HttpError, HttpResult, InfoHandler, MetricsBackendHandler, MetricsBlobcacheHandler,
-    MetricsFilesHandler, MetricsHandler, MetricsInflightHandler, MetricsPatternHandler,
-    MountHandler, SendFuseFdHandler, TakeoverHandler,
+    EventsHandler, ExitHandler, FsBackendInfo, InfoHandler, MetricsBackendHandler,
+    MetricsBlobcacheHandler, MetricsFilesHandler, MetricsHandler, MetricsInflightHandler,
+    MetricsPatternHandler, MountHandler, SendFuseFdHandler, TakeoverHandler, HTTP_ROOT_V1,
 };
 
-const HTTP_ROOT_V1: &str = "/api/v1";
 const EXIT_TOKEN: Token = Token(usize::MAX);
 const REQUEST_TOKEN: Token = Token(1);
+
+#[derive(Clone, Deserialize, Debug)]
+pub struct ApiMountCmd {
+    pub source: String,
+    #[serde(default)]
+    pub fs_type: String,
+    pub config: String,
+    #[serde(default)]
+    pub prefetch_files: Option<Vec<String>>,
+}
+
+#[derive(Clone, Deserialize, Debug)]
+pub struct ApiUmountCmd {
+    pub mountpoint: String,
+}
+
+#[derive(Clone, Deserialize, Debug)]
+pub struct DaemonConf {
+    pub log_level: String,
+}
+
+#[derive(Debug)]
+pub enum ApiRequest {
+    // Common requests
+    ConfigureDaemon(DaemonConf),
+
+    // Nydus API v1 requests
+    DaemonInfo,
+
+    Events,
+    Mount(String, ApiMountCmd),
+    Remount(String, ApiMountCmd),
+    Umount(String),
+    ExportGlobalMetrics(Option<String>),
+    ExportFilesMetrics(Option<String>, bool),
+    ExportAccessPatterns(Option<String>),
+    ExportBackendMetrics(Option<String>),
+    ExportBlobcacheMetrics(Option<String>),
+    ExportInflightMetrics,
+    ExportFsBackendInfo(String),
+    SendFuseFd,
+    Takeover,
+    Exit,
+}
+
+#[derive(Debug)]
+pub enum DaemonErrorKind {
+    NotReady,
+    UpgradeManager,
+    Unsupported,
+    Connect(io::Error),
+    SendFd,
+    RecvFd,
+    Disconnect(io::Error),
+    Channel,
+    Serde(SerdeError),
+    UnexpectedEvent(String),
+    Other(String),
+}
+
+#[derive(Debug)]
+pub enum MetricsErrorKind {
+    // Errors about daemon working status
+    Daemon(DaemonErrorKind),
+    // Errors about metrics/stats recorders
+    Stats(IoStatsError),
+}
+
+/// Errors generated by/related to the API service, sent back through `ApiResponse`.
+#[derive(Debug)]
+pub enum ApiError {
+    /// Daemon internal error
+    DaemonAbnormal(DaemonErrorKind),
+    /// Failed to get events information
+    Events(String),
+    /// Failed to get metrics information
+    Metrics(MetricsErrorKind),
+    /// Failed to mount filesystem
+    MountFilesystem(DaemonErrorKind),
+    /// Failed to send request to the API service
+    RequestSend(SendError<Option<ApiRequest>>),
+    /// Unrecognized payload content
+    ResponsePayloadType,
+    /// Failed to receive response from the API service
+    ResponseRecv(RecvError),
+    /// Failed to send wakeup notification
+    Wakeup(io::Error),
+}
+
+/// Specialized `std::result::Result` for API replies.
+pub type ApiResult<T> = std::result::Result<T, ApiError>;
+
+#[derive(Serialize)]
+pub enum ApiResponsePayload {
+    /// No data is sent on the channel.
+    Empty,
+    /// Daemon version, configuration and status information in json.
+    DaemonInfo(String),
+    Events(String),
+    FsBackendInfo(String),
+    /// Nydus filesystem global metrics
+    FsGlobalMetrics(String),
+    /// Nydus filesystem per-file metrics
+    FsFilesMetrics(String),
+    FsFilesPatterns(String),
+    BackendMetrics(String),
+    BlobcacheMetrics(String),
+    InflightMetrics(String),
+}
+
+/// HTTP error messages sent back to the clients.
+///
+/// The `HttpError` object will be sent back to client with `format!("{:?}", http_error)`.
+/// So unfortunately it implicitly becomes parts of the API, please keep it stable.
+#[derive(Debug)]
+pub enum HttpError {
+    /// No handler registered for HTTP request URI
+    NoRoute,
+    /// Invalid HTTP request
+    BadRequest,
+    /// Query parameter is missed from the HTTP request.
+    QueryString(String),
+    /// Failed to parse HTTP request message body
+    ParseBody(SerdeError),
+    /// Failed to query information about daemon.
+    DaemonInfo(ApiError),
+    Events(ApiError),
+    /// Could not mount resource
+    Mount(ApiError),
+    GlobalMetrics(ApiError),
+    FsFilesMetrics(ApiError),
+    Pattern(ApiError),
+    Configure(ApiError),
+    Upgrade(ApiError),
+    BlobcacheMetrics(ApiError),
+    BackendMetrics(ApiError),
+    FsBackendInfo(ApiError),
+    InflightMetrics(ApiError),
+}
+
+/// This is the response sent by the API server through the mpsc channel.
+pub type ApiResponse = std::result::Result<ApiResponsePayload, ApiError>;
+pub type HttpResult = std::result::Result<Response, HttpError>;
+
+#[derive(Serialize, Debug)]
+struct ErrorMessage {
+    code: String,
+    message: String,
+}
+
+impl From<ErrorMessage> for Vec<u8> {
+    fn from(msg: ErrorMessage) -> Self {
+        // Safe to unwrap since `ErrorMessage` must succeed in serialization
+        serde_json::to_vec(&msg).unwrap()
+    }
+}
+
+pub fn extract_query_part(req: &Request, key: &str) -> Option<String> {
+    // Splicing req.uri with "http:" prefix might look weird, but since it depends on
+    // crate `Url` to generate query_pairs HashMap, which is working on top of Url not Uri.
+    // Better that we can add query part support to Micro-http in the future. But
+    // right now, below way makes it easy to obtain query parts from uri.
+    let http_prefix = format!("http:{}", req.uri().get_abs_path());
+    let url = Url::parse(&http_prefix)
+        .map_err(|e| {
+            error!("Can't parse request {:?}", e);
+            e
+        })
+        .ok()?;
+
+    for (k, v) in url.query_pairs() {
+        if k == key {
+            trace!("Got query part {:?}", (k, &v));
+            return Some(v.into_owned());
+        }
+    }
+    None
+}
+
+/// Parse HTTP request body.
+pub(crate) fn parse_body<'a, F: Deserialize<'a>>(b: &'a Body) -> std::result::Result<F, HttpError> {
+    serde_json::from_slice::<F>(b.raw()).map_err(HttpError::ParseBody)
+}
+
+/// Translate ApiError message to HTTP status code.
+pub(crate) fn translate_status_code(e: &ApiError) -> StatusCode {
+    match e {
+        ApiError::DaemonAbnormal(kind) | ApiError::MountFilesystem(kind) => match kind {
+            DaemonErrorKind::NotReady => StatusCode::ServiceUnavailable,
+            DaemonErrorKind::Unsupported => StatusCode::NotImplemented,
+            DaemonErrorKind::UnexpectedEvent(_) => StatusCode::BadRequest,
+            _ => StatusCode::InternalServerError,
+        },
+        ApiError::Metrics(MetricsErrorKind::Stats(IoStatsError::NoCounter)) => StatusCode::NotFound,
+        _ => StatusCode::InternalServerError,
+    }
+}
+
+/// Generate a successful HTTP response message.
+pub(crate) fn success_response(body: Option<String>) -> Response {
+    if let Some(body) = body {
+        let mut r = Response::new(Version::Http11, StatusCode::OK);
+        r.set_body(Body::new(body));
+        r
+    } else {
+        Response::new(Version::Http11, StatusCode::NoContent)
+    }
+}
+
+/// Generate a HTTP error response message with status code and error message.
+pub(crate) fn error_response(error: HttpError, status: StatusCode) -> Response {
+    let mut response = Response::new(Version::Http11, status);
+    let err_msg = ErrorMessage {
+        code: "UNDEFINED".to_string(),
+        message: format!("{:?}", error),
+    };
+    response.set_body(Body::new(err_msg));
+    response
+}
 
 /// Trait for HTTP endpoints to handle HTTP requests.
 pub trait EndpointHandler: Sync + Send {
@@ -92,7 +311,7 @@ fn kick_api_server(
 ) -> ApiResponse {
     to_api.send(Some(request)).map_err(ApiError::RequestSend)?;
     if let Some(waker) = api_notifier {
-        waker.wake().map_err(ApiError::EventFdWrite)?;
+        waker.wake().map_err(ApiError::Wakeup)?;
     }
     from_api.recv().map_err(ApiError::ResponseRecv)?
 }
@@ -247,28 +466,6 @@ pub fn start_http_thread(
         })?;
 
     Ok((thread, waker))
-}
-
-pub fn extract_query_part(req: &Request, key: &str) -> Option<String> {
-    // Splicing req.uri with "http:" prefix might look weird, but since it depends on
-    // crate `Url` to generate query_pairs HashMap, which is working on top of Url not Uri.
-    // Better that we can add query part support to Micro-http in the future. But
-    // right now, below way makes it easy to obtain query parts from uri.
-    let http_prefix = format!("http:{}", req.uri().get_abs_path());
-    let url = Url::parse(&http_prefix)
-        .map_err(|e| {
-            error!("Can't parse request {:?}", e);
-            e
-        })
-        .ok()?;
-
-    for (k, v) in url.query_pairs() {
-        if k == key {
-            trace!("Got query part {:?}", (k, &v));
-            return Some(v.into_owned());
-        }
-    }
-    None
 }
 
 #[cfg(test)]
